[FIX] delete file

+++ utils/deleteFile.js
@@ -6,30 +6,33 @@ const deleteFile = async (url) => {
   try {
     if (process.env.APP_ENV === "local") {
       const filename = url.split(
-        `http://localhost:${process.env.PORT}/${process.env.LOCAL_UPLOAD_FOLDER}/`
+        `http://localhost:${process.env.PORT}/api/v1/file/`
       );
-
-      const filePath = path.join(
-        path.resolve(__dirname, ".."),
-        process.env.LOCAL_UPLOAD_FOLDER,
-        filename[1]
-      );
-
-      if (fs.existsSync(filePath)) {
-        fs.unlink(filePath, (err) => {
-          if (err) {
-            return false;
-          }
-        });
-        return true;
-      } else {
-        return false;
+      
+      if(filename[1]){
+        const filePath = path.join(
+          path.resolve(__dirname, ".."),
+          process.env.LOCAL_UPLOAD_FOLDER,
+          filename[1]
+        );
+  
+        if (fs.existsSync(filePath)) {
+          fs.unlink(filePath, (err) => {
+            if (err) {
+              return false;
+            }
+          });
+          return true;
+        }
       }
+      return true;  
     } else {
       await del(url);
       return true;
     }
   } catch (error) {
+    console.log(error);
+    
     return false;
   }
 };
